pyproject.toml
[tool.poetry]
name = "doreumung-project"
version = "0.1.0"
description = ""
authors = ["limsu1015 <tngur1015@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.34.0"
psycopg2 = "^2.9.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.4"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.0"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.0"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
beautifulsoup4 = "^4.12.3"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
psycopg2-binary = "^2.9.10"
greenlet = "^3.1.1"
httpx = "^0.28.1"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.25.0"
respx = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 88
known_third_party = ["fastapi", "pydantic", "sqlalchemy", "uvicorn"]

[tool.mypy]
strict = true
ignore_missing_imports = true